% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_afstemningsomraader.R
\docType{data}
\name{geo_afstemningsomraader}
\alias{geo_afstemningsomraader}
\title{Data from the DAGI API}
\format{
A data frame with ? rows and ? variables:
\describe{
  \item{afstemningssted_adgangspunkt_x}{Individuel kode for kommunen}
  \item{afstemningssted_adgangspunkt_y}{Individuel kode for kommunen}
  \item{afstemningsstedadressebetegnelse}{Individuel kode for kommunen}
  \item{afstemningsstedadresseid}{Individuel kode for kommunen}
  \item{afstemningsstednavn}{Individuel kode for kommunen}
  \item{changed}{Individuel kode for kommunen}
  \item{dagi_id}{Individuel kode for kommunen}
  \item{geo_changed}{Individuel kode for kommunen}
  \item{geo_version}{Individuel kode for kommunen}
  \item{geometry}{Individuel kode for kommunen}
  \item{kommunekode}{Individuel kode for kommunen}
  \item{kommunenavn}{Individuel kode for kommunen}
  \item{navn}{Individuel kode for kommunen}
  \item{nummer}{Individuel kode for kommunen}
  \item{opstillingskredsnavn}{Individuel kode for kommunen}
  \item{opstillingskredsnummer}{Individuel kode for kommunen}
  \item{regionskode}{Individuel kode for kommunen}
  \item{regionsnavn}{Individuel kode for kommunen}
  \item{storkredsnavn}{Individuel kode for kommunen}
  \item{storkredsnummer}{Individuel kode for kommunen}
  \item{valglandsdelsbogstav}{Individuel kode for kommunen}
  \item{valglandsdelsnavn}{Individuel kode for kommunen}
  \item{visueltcenter_x}{Individuel kode for kommunen}
  \item{visueltcenter_y}{Individuel kode for kommunen}
}
}
\source{
\href{http://dawa.dk}{Website}
}
\usage{
data(geo_afstemningsomraader)
}
\description{
Data from the DAGI API
}
\examples{
data(geo_afstemningsomraader)


}
\keyword{datasets}
