% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_f07.R
\docType{data}
\name{geo_f07}
\alias{geo_f07}
\title{Data from the DAGI API}
\format{
A data frame with ? rows and ? variables:
\describe{
  \item{dagi_id}{Individuel kode for kommunen}
  \item{kode}{Individuel kode for kommunen}
  \item{navn}{Individuel kode for kommunen}
  \item{regionskode}{Individuel kode for kommunen}
  \item{udenforkommuneinddeling}{Individuel kode for kommunen}
  \item{regionsnavn}{Individuel kode for kommunen}
  \item{changed}{Individuel kode for kommunen}
  \item{geo_changed}{Individuel kode for kommunen}
  \item{geo_version}{Individuel kode for kommunen}
  \item{visueltcenter_x}{Individuel kode for kommunen}
  \item{visueltcenter_y}{Individuel kode for kommunen}
  \item{area}{Individuel kode for kommunen}
  \item{population}{Individuel kode for kommunen}
  \item{men}{Individuel kode for kommunen}
  \item{women}{Individuel kode for kommunen}
  \item{pct_men}{Individuel kode for kommunen}
  \item{avg_age}{Individuel kode for kommunen}
  \item{stat_year}{Individuel kode for kommunen}
  \item{geometry }{Individuel kode for kommunen}
}
}
\source{
\href{http://??.dk}{Website}
}
\usage{
data(geo_f07)
}
\description{
Data from the DAGI API
}
\examples{
data(geo_f07)


}
\keyword{datasets}
