% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_opstillingskredse.R
\docType{data}
\name{geo_opstillingskredse}
\alias{geo_opstillingskredse}
\title{Data from the DAGI API}
\format{
A data frame with ? rows and ? variables:
\describe{
  \item{dagi_id}{Individuel kode for kommunen}
  \item{nummer}{Individuel kode for kommunen}
  \item{kode}{Individuel kode for kommunen}
  \item{navn}{Individuel kode for kommunen}
  \item{regionskode}{Individuel kode for kommunen}
  \item{regionsnavn}{Individuel kode for kommunen}
  \item{kredskommunekode}{Individuel kode for kommunen}
  \item{kredskommunenavn}{Individuel kode for kommunen}
  \item{storkredsnummer}{Individuel kode for kommunen}
  \item{storkredsnavn}{Individuel kode for kommunen}
  \item{valglandsdelsbogstav}{Individuel kode for kommunen}
  \item{valglandsdelsnavn}{Individuel kode for kommunen}
  \item{valgkredsnummer}{Individuel kode for kommunen}
  \item{changed}{Individuel kode for kommunen}
  \item{geo_changed}{Individuel kode for kommunen}
  \item{geo_version}{Individuel kode for kommunen}
  \item{visueltcenter_x}{Individuel kode for kommunen}
  \item{visueltcenter_y}{Individuel kode for kommunen}
  \item{geometry}{Individuel kode for kommunen}
}
}
\source{
\href{http://dawa.dk}{Website}
}
\usage{
data(geo_opstillingskredse)
}
\description{
Data from the DAGI API
}
\examples{
data(geo_opstillingskredse)


}
\keyword{datasets}
