% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_politikredse.R
\docType{data}
\name{geo_politikredse}
\alias{geo_politikredse}
\title{Data from the DAGI API}
\format{
A data frame with ? rows and ? variables:
\describe{
  \item{dagi_id}{Individuel kode for kommunen}
  \item{kode}{Individuel kode for kommunen}
  \item{navn}{Individuel kode for kommunen}
  \item{changed}{Individuel kode for kommunen}
  \item{geo_changed}{Individuel kode for kommunen}
  \item{geo_version}{Individuel kode for kommunen}
  \item{visueltcenter_x}{Individuel kode for kommunen}
  \item{visueltcenter_y}{Individuel kode for kommunen}
  \item{geometry}{Individuel kode for kommunen}
}
}
\source{
\href{http://dawa.dk}{Website}
}
\usage{
data(geo_politikredse)
}
\description{
Data from the DAGI API
}
\examples{
data(geo_politikredse)


}
\keyword{datasets}
