% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kom_tabel.R
\docType{data}
\name{kom_tabel}
\alias{kom_tabel}
\title{Data from the DST}
\format{
A data frame with ? rows and ? variables:
\describe{
  \item{KØN}{Individuel kode for kommunen}
  \item{ALDER}{Individuel kode for kommunen}
  \item{INDHOLD}{Individuel kode for kommunen}
  \item{area}{Individuel kode for kommunen}
}
}
\source{
\href{http://dawa.dk}{Website}
}
\usage{
data(kom_tabel)
}
\description{
Data from the DST
}
\examples{
data(kom_tabel)


}
\keyword{datasets}
