% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/std_kommuner.R
\docType{data}
\name{std_kommuner}
\alias{std_kommuner}
\title{Data from the DST}
\format{
A data frame with ? rows and ? variables:
\describe{
  \item{KØN}{Individuel kode for kommunen}
  \item{ALDER}{Individuel kode for kommunen}
  \item{INDHOLD}{Individuel kode for kommunen}
  \item{area}{Individuel kode for kommunen}
}
}
\source{
\href{http://dawa.dk}{Website}
}
\usage{
data(std_kommuner)
}
\description{
Data from the DST
}
\examples{
data(std_kommuner)


}
\keyword{datasets}
