% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/std_sogne.R
\docType{data}
\name{std_sogne}
\alias{std_sogne}
\title{Data from the DAGI API}
\format{
A data frame with ? rows and ? variables:
\describe{
  \item{KØN}{Individuel kode for kommunen}
  \item{ALDER}{Individuel kode for kommunen}
  \item{INDHOLD}{Individuel kode for kommunen}
  \item{kode}{Individuel kode for kommunen}
}
}
\source{
\href{http://dawa.dk}{Website}
}
\usage{
data(std_sogne)
}
\description{
Data from the DAGI API
}
\examples{
data(std_sogne)


}
\keyword{datasets}
