% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_landsdele.R
\docType{data}
\name{geo_landsdele}
\alias{geo_landsdele}
\title{Data from the DAGI API}
\format{
A data frame with ? rows and ? variables:
\describe{
  \item{avg_age}{Individuel kode for kommunen}
  \item{dagi_id}{Individuel kode for kommunen}
  \item{geo_version}{Individuel kode for kommunen}
  \item{geo_changed}{Individuel kode for kommunen}
  \item{changed}{Individuel kode for kommunen}
  \item{eometry }{Individuel kode for kommunen}
  \item{kode}{Individuel kode for kommunen}
  \item{men}{Individuel kode for kommunen}
  \item{women}{Individuel kode for kommunen}
  \item{navn}{Individuel kode for kommunen}
  \item{nuts2}{Individuel kode for kommunen}
  \item{pct_men}{Individuel kode for kommunen}
  \item{population}{Individuel kode for kommunen}
  \item{stat_year}{Individuel kode for kommunen}
  \item{visueltcenter_x}{Individuel kode for kommunen}
  \item{visueltcenter_y}{Individuel kode for kommunen}
}
}
\source{
\href{http://dawa.dk}{Website}
}
\usage{
data(geo_landsdele)
}
\description{
Data from the DAGI API
}
\examples{
data(geo_landsdele)


}
\keyword{datasets}
